% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avert.R
\name{avert}
\alias{avert}
\title{Title}
\usage{
avert(
  hourly_load_reduction,
  project_year,
  project_region,
  avertr_rdf_filepath,
  avert_main_module_filepath
)
}
\arguments{
\item{project_region}{}
}
\description{
Title
}
