% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_and_avert.R
\name{generate_and_avert}
\alias{generate_and_avert}
\title{Generate an hourly load reduction vector and run an AVERT scenario}
\usage{
generate_and_avert(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[=generate_reduction]{generate_reduction()}} (and then to
\code{\link[=avert]{avert()}}).}
}
\value{
A list with three elements:
\enumerate{
\item \code{differences_final}, a tibble. In a region with N fossil-fuel generating
units, the tibble will have N * 8760 rows. Each row contains the changes
in emissions (and generation and heat input) associated with a given
fossil-fuel generating unit in a given hour. Those changes appear in
the columns with a "data_" prefix. The other columns given additional
information on the generating unit and the fossil-fuel load bin.
\item \code{signal_to_noise}, a two-element list containing A. the linear
regression coefficients which result from regressing the generation
change calculated by avertr onto the input generation change, and
B. the r squared statistic from same regression.
\item \code{pct_hourly_load_reduction}, an 8760-length (or, in a leap year,
8784-length) numeric vector where each value represents the percent
reduction in generation compared to the business-as-usual scenario.
Presented as numbers ranging from 0 to 1. E.g., 0.028 would represent
a 2.8\% decrease in generation. AVERT suggests caution when modeling
a scenario where any given hour has a >15\% change in generation.
}
}
\description{
\code{generate_and_avert()} is a wrapper function that generates an hourly
load reduction vector with \code{\link[=generate_reduction]{generate_reduction()}} and then runs
an AVERT scenario with \code{\link[=avert]{avert()}}, all in one step. It's more convenient
than running the two functions separately, especially because there's no
need to re-specify the shared arguments (\code{project_year}, \code{project_region},
\code{avert_main_module_filepath}, and \code{avertr_rdf_filepath}).
}
\examples{
\dontrun{
# In two steps with generate_reduction() and avert()
generate_reduction(
  offshore_wind_capacity_mw = 200,
  apply_reduction_top_x_pct_hours = 5,
  reduce_x_pct_in_top_hours = 10,
  project_year = 2023,
  project_region = "New England",
  avert_main_module_filepath = "./avert-main-module-v4.3.xlsx",
  avertr_rdf_filepath = "./avertr_rdfs/2023/avertr_rdf_New England_2023.rds"
) |>
avert(
  project_year = 2023,
  project_region = "New England",
  avert_main_module_filepath = "./avert-main-module-v4.3.xlsx",
  avertr_rdf_filepath = "./avertr_rdfs/2023/avertr_rdf_New England_2023.rds"
)

# In one step with generate_and_avert()
generate_and_avert(
  offshore_wind_capacity_mw = 200,
  apply_reduction_top_x_pct_hours = 5,
  reduce_x_pct_in_top_hours = 10,
  project_year = 2023,
  project_region = "New England",
  avert_main_module_filepath = "./avert-main-module-v4.3.xlsx",
  avertr_rdf_filepath = "./avertr_rdfs/2023/avertr_rdf_New England_2023.rds"
)
}
}
