% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avertr.R
\name{adjust_reduction}
\alias{adjust_reduction}
\title{Title}
\usage{
adjust_reduction(
  unadjusted_hourly_load_reduction = NULL,
  project_year = NULL,
  project_region = NULL
)
}
\arguments{
\item{project_region}{}
}
\description{
Title
}
