% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avertr.R
\name{generate_reduction}
\alias{generate_reduction}
\title{Title}
\usage{
generate_reduction(
  project_year,
  project_region,
  avert_main_module_filepath,
  avertr_rdf_filepath,
  apply_reduction_top_x_pct_hours = 0,
  reduce_x_pct_in_top_hours = 0,
  reduce_annual_generation_by_x_gwh = 0,
  reduce_each_hour_by_x_mw = 0,
  onshore_wind_capacity_mw = 0,
  offshore_wind_capacity_mw = 0,
  utility_solar_pv_capacity_mw = 0,
  rooftop_solar_pv_capacity_mw = 0
)
}
\arguments{
\item{rooftop_solar_pv_capacity_mw}{}
}
\description{
Title
}
