% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avertr_test.R
\name{test_hourly}
\alias{test_hourly}
\title{Test avertr results against AVERT results (hourly level)}
\usage{
test_hourly(avertr_results, avert_run_filepath)
}
\arguments{
\item{avertr_results}{The output from an \code{\link[=avert]{avert()}} run.}

\item{avert_run_filepath}{A string giving a filepath to a finalized AVERT
run which has been saved as a .xlsx file.}
}
\value{
A list with at least two (possibly three) elements:
\enumerate{
\item \code{error_summary_table_hourly}, a table giving summaries of the
difference and percent difference between the avertr results and AVERT
where the summaries range across different generating unit-hours. A
generating unit-hour is a given generating unit in a given hour. So
if there are N generating units there will be 8760 * N generating-unit
hours (in a non-leap year). E.g., if the "Mean" value for pm25_error
is -0.0004137, that means that across all generating unit-hours in the
region, the average difference between that unit's PM2.5 change in
avertr and its PM2.5 change in AVERT is -0.0004137 (lbs).
\item \code{largest_absolute_errors_hourly}, a one-row tibble where each value is
the largest absolute value error (or absolute value percent error)
across all generating units-hours in the region. E.g., if pct_nox_error
is 0.12, that  means that the generating unit-hour with the largest
absolute value percent error between avertr's and AVERT's reported NOx
changes had a NOx value that was 0.12\% different between avertr and AVERT.
\item (Possibly) \code{absolute_pct_errors_above_001_hourly}, a tibble containing all
generating unit-hours which have an absolute value percent error greater
than 0.1\% for at least one of their "\code{data_}" columns. If there are no
such generating unit-hours, this table is not returned, and the list
only has two elements.
}
}
\description{
\code{test_hourly()} takes an avertr run and an AVERT run and returns information
about the differences between the two at the hourly level. It can be used to
verify that, with the same inputs, avertr and AVERT produce results which
differ only negligibly.
}
\details{
\code{test_hourly()} checks for differences at the hourly level, meaning that checks
for differences between avertr and AVERT in each hour of the year, for each
generating unit.
}
\examples{
\dontrun{
avert_out <- avert(
  hourly_load_reduction = rep(100, 8760),
  project_year = 2021,
  project_region = "Florida",
  avert_main_module_filepath = "./avert-main-module-v4.3.xlsx",
  avertr_rdf_filepath = "./avertr_rdfs/2021/avertr_rdf_Florida_2021.rds"
)

test_hourly(avert_out, "./test_scenarios/100MW_flat_reduction_Florida_2021.xlsx")
}
}
